fix clock bug

+++ inc/write_data.f90
@@ -27,7 +27,7 @@ character(len=64):: title
 !            steps*steps/2 total iterations
 real(dp)         :: qstep
 integer          :: steps, i, iqx, iqy, log=10, dat=20
-real             :: start, end, rate, est
+integer          :: start, end, rate, est
 
 ! Variables for dcuhre
 !     ndim   integer.
